GADAM/requirements.txt
torch==1.12.0
torchvision==0.13.0
torchaudio==0.12.0
dgl-cu116
scikit-learn
pygod
numpy
scipy
pytorch-memlab 